Rewrite DistributedExt once more with DistributedArrays
This approach seems much more reasonable and seems to have no memory
leaks.

We centrally allocate all the buffers, however we do it through
DistributedArrays which correctly can move the data to the workers.

We need to make sure to use the `localpart` command in the submethods
when accessing the buffer, which is a no-op if the buffer is already
local, and otherwise get's the correct buffer.