Bring back `minibatch`
We had previously removed minibatching because it's a bit tricky to get
to work with Matryoshka. However, we managed to bring it back.
To this end, we also introduced `maybeview`: A function that constructs
a view if the column indexing is contiguous, and otherwise copies.

The new minibatching continues to only update `D` on some subset of the
columns of `Y`, but returns `X` for all columns of `Y`, as it's good to
update the structure.